# Add an optional ability to skip comments