--- a/jsoniter-scala-core/jvm/src/main/java/com/github/plokhotnyuk/jsoniter_scala/core/ByteArrayAccess.java
+++ b/jsoniter-scala-core/jvm/src/main/java/com/github/plokhotnyuk/jsoniter_scala/core/ByteArrayAccess.java
@@ -36,6 +36,10 @@ class ByteArrayAccess { // FIXME: Use Java wrapper as w/a for missing support of
         VH_SHORT.set(buf, pos, value);
     }
 
+    static short getShort(byte[] buf, int pos) {
+        return (short) VH_SHORT.get(buf, pos);
+    }
+
     static void setLongReversed(byte[] buf, int pos, long value) {
         VH_LONG_REVERSED.set(buf, pos, value);
     }