--- a/jsoniter-scala-core/jvm/src/main/scala/com/github/plokhotnyuk/jsoniter_scala/core/JsonReader.scala
+++ b/jsoniter-scala-core/jvm/src/main/scala/com/github/plokhotnyuk/jsoniter_scala/core/JsonReader.scala
@@ -549,9 +549,13 @@ final class JsonReader private[jsoniter_scala](
     else if ((b >= '0' && b <= '9') || b == '-') skipNumber(head)
     else if (b == 'n' || b == 't') skipFixedBytes(3, head)
     else if (b == 'f') skipFixedBytes(4, head)
-    else if (b == '[') skipArray(0, head)
-    else if (b == '{') skipObject(0, head)
-    else decodeError("expected value")
+    else if (b == '[') {
+      if (config.allowComments) skipArrayWithComments(0, head)
+      else skipArray(0, head)
+    } else if (b == '{') {
+      if (config.allowComments) skipObjectWithComments(0, head)
+      else skipObject(0, head)
+    } else decodeError("expected value")
   }
 
   def commaError(): Nothing = decodeError("expected ','")
@@ -721,7 +725,10 @@ final class JsonReader private[jsoniter_scala](
       pos < tail
     }) && {
       val b = buf(pos)
-      b == ' ' || b == '\n' || (b | 0x4) == '\r'
+      b == ' ' || b == '\n' || (b | 0x4) == '\r' || b == '/' && config.allowComments && {
+        pos = skipComment(pos + 1) - 1
+        true
+      }
     }) pos += 1
     head = pos
     pos != tail
@@ -792,6 +799,7 @@ final class JsonReader private[jsoniter_scala](
     if (pos < tail) {
       val b = buf(pos)
       if (b == ' ' || b == '\n' || (b | 0x4) == '\r') nextToken(pos + 1)
+      else if (b == '/' && config.allowComments) nextToken(skipComment(pos + 1))
       else {
         head = pos + 1
         b
@@ -803,7 +811,7 @@ final class JsonReader private[jsoniter_scala](
     if (pos < tail) {
       val b = buf(pos)
       head = pos + 1
-      if (b != t && ((b != ' ' && b != '\n' && (b | 0x4) != '\r') || nextToken(pos + 1) != t)) tokenError(t, head - 1)
+      if (b != t && nextToken(pos) != t) tokenError(t, head - 1)
     } else nextTokenOrError(t, loadMoreOrError(pos))
 
   @tailrec
@@ -811,7 +819,7 @@ final class JsonReader private[jsoniter_scala](
     if (pos < tail) {
       val b = buf(pos)
       head = pos + 1
-      b == t || ((b == ' ' || b == '\n' || (b | 0x4) == '\r') && nextToken(pos + 1) == t)
+      b == t || nextToken(pos) == t
     } else isNextToken(t, loadMoreOrError(pos))
 
   private[this] def isCurrentToken(t: Byte, pos: Int): Boolean = {
@@ -1168,11 +1176,12 @@ final class JsonReader private[jsoniter_scala](
       } else if (bs == 0x736C6166) {
         if (nextByte(pos + 4) != 'e') booleanError(pos + 4)
         false
-      } else if (isToken && {
-        val b1 = bs.toByte
-        b1 == ' ' || b1 == '\n' || (b1 | 0x4) == '\r'
-      }) parseBoolean(isToken, pos + 1)
-      else booleanError(bs, pos)
+      } else if (isToken) {
+        val b = bs.toByte
+        if (b == ' ' || b == '\n' || (b | 0x4) == '\r') parseBoolean(isToken, pos + 1)
+        else if (b == '/' && config.allowComments) parseBoolean(isToken, skipComment(pos + 1))
+        else booleanError(bs, pos)
+      } else booleanError(bs, pos)
     } else parseBoolean(isToken, loadMoreOrError(pos))
 
   private[this] def booleanError(bs: Int, pos: Int): Nothing = booleanError({
@@ -3448,10 +3457,11 @@ final class JsonReader private[jsoniter_scala](
     if (pos < tail) {
       val b = buf(pos)
       if (b == '"') skipObject(level, skipString(evenBackSlashes = true, pos + 1))
-      else if (b == '{') skipObject(level + 1, pos + 1)
-      else if (b != '}') skipObject(level, pos + 1)
-      else if (level != 0) skipObject(level - 1, pos + 1)
-      else pos + 1
+      else if (b == '}') {
+        if (level == 0) pos + 1
+        else skipObject(level - 1, pos + 1)
+      } else if (b == '{') skipObject(level + 1, pos + 1)
+      else skipObject(level, pos + 1)
     } else skipObject(level, loadMoreOrError(pos))
 
   @tailrec
@@ -3459,11 +3469,61 @@ final class JsonReader private[jsoniter_scala](
     if (pos < tail) {
       val b = buf(pos)
       if (b == '"') skipArray(level, skipString(evenBackSlashes = true, pos + 1))
-      else if (b == '[') skipArray(level + 1, pos + 1)
-      else if (b != ']') skipArray(level, pos + 1)
-      else if (level != 0) skipArray(level - 1, pos + 1)
-      else pos + 1
+      else if (b == ']') {
+        if (level == 0) pos + 1
+        else skipArray(level - 1, pos + 1)
+      } else if (b == '[') skipArray(level + 1, pos + 1)
+      else skipArray(level, pos + 1)
     } else skipArray(level, loadMoreOrError(pos))
+
+  @tailrec
+  private[this] def skipObjectWithComments(level: Int, pos: Int): Int =
+    if (pos < tail) {
+      val b = buf(pos)
+      if (b == '"') skipObjectWithComments(level, skipString(evenBackSlashes = true, pos + 1))
+      else if (b == '}') {
+        if (level == 0) pos + 1
+        else skipObjectWithComments(level - 1, pos + 1)
+      } else if (b == '{') skipObjectWithComments(level + 1, pos + 1)
+      else if (b == '/') skipObjectWithComments(level, skipComment(pos + 1))
+      else skipObjectWithComments(level, pos + 1)
+    } else skipObjectWithComments(level, loadMoreOrError(pos))
+
+  @tailrec
+  private[this] def skipArrayWithComments(level: Int, pos: Int): Int =
+    if (pos < tail) {
+      val b = buf(pos)
+      if (b == '"') skipArrayWithComments(level, skipString(evenBackSlashes = true, pos + 1))
+      else if (b == ']') {
+        if (level == 0) pos + 1
+        else skipArrayWithComments(level - 1, pos + 1)
+      } else if (b == '[') skipArrayWithComments(level + 1, pos + 1)
+      else if (b == '/') skipArrayWithComments(level, skipComment(pos + 1))
+      else skipArrayWithComments(level, pos + 1)
+    } else skipArrayWithComments(level, loadMoreOrError(pos))
+
+  @tailrec
+  private[this] def skipComment(pos: Int): Int =
+    if (pos < tail) {
+      val b = buf(pos)
+      if (b == '/') skipOneLineComment(pos + 1)
+      else if (b == '*') skipMultiLineComment(pos + 1)
+      else tokensError('/', '*', pos)
+    } else skipComment(loadMoreOrError(pos))
+
+  @tailrec
+  private[this] def skipOneLineComment(pos: Int): Int =
+    if (pos < tail) {
+      if (buf(pos) != '\n') skipOneLineComment(pos + 1)
+      else pos + 1
+    } else skipOneLineComment(loadMoreOrError(pos))
+
+  @tailrec
+  private[this] def skipMultiLineComment(pos: Int): Int =
+    if (pos + 1 < tail) {
+      if (ByteArrayAccess.getShort(buf, pos) != 0x2F2A) skipMultiLineComment(pos + 1)
+      else pos + 2
+    } else skipMultiLineComment(loadMoreOrError(pos))
 
   @tailrec
   private[this] def skipFixedBytes(n: Int, pos: Int): Int = {