Throughput of parsing with Zulu 17

## Before
```
[info] Benchmark                                               (size)   Mode  Cnt         Score        Error   Units
[info] ADTReading.jsoniterScala                                   N/A  thrpt    5   5838254.845 ±  82581.432   ops/s
[info] AnyValsReading.jsoniterScala                               N/A  thrpt    5   8902001.249 ±  77923.479   ops/s
[info] ArrayBufferOfBooleansReading.jsoniterScala                 128  thrpt    5   2043279.701 ±  48458.294   ops/s
[info] ArrayOfBigDecimalsReading.jsoniterScala                    128  thrpt    5    145220.747 ±    657.771   ops/s
[info] ArrayOfBigIntsReading.jsoniterScala                        128  thrpt    5    227700.054 ±   3370.713   ops/s
[info] ArrayOfBooleansReading.jsoniterScala                       128  thrpt    5   2626008.631 ±  65010.511   ops/s
[info] ArrayOfBytesReading.jsoniterScala                          128  thrpt    5   1714490.536 ±  15383.768   ops/s
[info] ArrayOfCharsReading.jsoniterScala                          128  thrpt    5   1848570.468 ±  47932.488   ops/s
[info] ArrayOfDoublesReading.jsoniterScala                        128  thrpt    5    390446.982 ±   3221.216   ops/s
[info] ArrayOfDurationsReading.jsoniterScala                      128  thrpt    5    280538.524 ±   9880.198   ops/s
[info] ArrayOfEnumADTsReading.jsoniterScala                       128  thrpt    5    784598.372 ±  15261.733   ops/s
[info] ArrayOfEnumsReading.jsoniterScala                          128  thrpt    5    414009.422 ±   2953.813   ops/s
[info] ArrayOfFloatsReading.jsoniterScala                         128  thrpt    5    542573.775 ±  32585.431   ops/s
[info] ArrayOfInstantsReading.jsoniterScala                       128  thrpt    5    437053.970 ±  19668.575   ops/s
[info] ArrayOfIntsReading.jsoniterScala                           128  thrpt    5   1006065.699 ±   5411.792   ops/s
[info] ArrayOfJavaEnumsReading.jsoniterScala                      128  thrpt    5    838367.979 ±   9132.671   ops/s
[info] ArrayOfLocalDateTimesReading.jsoniterScala                 128  thrpt    5    432353.945 ±  13039.243   ops/s
[info] ArrayOfLocalDatesReading.jsoniterScala                     128  thrpt    5    946205.338 ±  56606.333   ops/s
[info] ArrayOfLocalTimesReading.jsoniterScala                     128  thrpt    5    659429.451 ±   8660.142   ops/s
[info] ArrayOfLongsReading.jsoniterScala                          128  thrpt    5    655863.312 ±  12942.692   ops/s
[info] ArrayOfMonthDaysReading.jsoniterScala                      128  thrpt    5   1260573.440 ±  19486.110   ops/s
[info] ArrayOfOffsetDateTimesReading.jsoniterScala                128  thrpt    5    371107.446 ±   4972.295   ops/s
[info] ArrayOfOffsetTimesReading.jsoniterScala                    128  thrpt    5    504312.482 ±   3287.725   ops/s
[info] ArrayOfPeriodsReading.jsoniterScala                        128  thrpt    5    325271.888 ±  11641.394   ops/s
[info] ArrayOfShortsReading.jsoniterScala                         128  thrpt    5   1475340.626 ±  10143.470   ops/s
[info] ArrayOfUUIDsReading.jsoniterScala                          128  thrpt    5    571735.112 ±   9234.696   ops/s
[info] ArrayOfYearMonthsReading.jsoniterScala                     128  thrpt    5   1457355.387 ±   8629.311   ops/s
[info] ArrayOfYearsReading.jsoniterScala                          128  thrpt    5   1702012.502 ±  48309.547   ops/s
[info] ArrayOfZoneIdsReading.jsoniterScala                        128  thrpt    5    394645.230 ±  13163.507   ops/s
[info] ArrayOfZoneOffsetsReading.jsoniterScala                    128  thrpt    5   1398263.414 ±  16157.069   ops/s
[info] ArrayOfZonedDateTimesReading.jsoniterScala                 128  thrpt    5    153648.030 ±   2513.474   ops/s
[info] ArraySeqOfBooleansReading.jsoniterScala                    128  thrpt    5   2610933.363 ±  43704.805   ops/s
[info] Base16Reading.jsoniterScala                                128  thrpt    5   8871585.392 ± 114876.198   ops/s
[info] Base64Reading.jsoniterScala                                128  thrpt    5   6494143.562 ± 163408.876   ops/s
[info] BigDecimalReading.jsoniterScala                            128  thrpt    5   6850270.879 ±  95210.586   ops/s
[info] BigIntReading.jsoniterScala                                128  thrpt    5   6756491.191 ± 183275.044   ops/s
[info] BitSetReading.jsoniterScala                                128  thrpt    5   1818506.133 ±  78906.477   ops/s
[info] ExtractFieldsReading.jsoniterScala                         128  thrpt    5    252291.591 ±   1036.051   ops/s
[info] GeoJSONReading.jsoniterScala                               N/A  thrpt    5     71415.145 ±    509.459   ops/s
[info] GitHubActionsAPIReading.jsoniterScala                      N/A  thrpt    5    771116.036 ±   6044.520   ops/s
[info] GoogleMapsAPIReading.jsoniterScala                         N/A  thrpt    5     36262.653 ±    144.401   ops/s
[info] IntMapOfBooleansReading.jsoniterScala                      128  thrpt    5    264945.490 ±  19575.321   ops/s
[info] IntReading.jsoniterScala                                   N/A  thrpt    5  86674670.478 ± 327517.377   ops/s
[info] ListOfBooleansReading.jsoniterScala                        128  thrpt    5   2375661.382 ±  44846.864   ops/s
[info] MapOfIntsToBooleansReading.jsoniterScala                   128  thrpt    5    212510.751 ±   2307.272   ops/s
[info] MissingRequiredFieldsReading.jsoniterScala                 N/A  thrpt    5   6179343.979 ±  93491.269   ops/s
[info] MissingRequiredFieldsReading.jsoniterScalaWithStacktrace   N/A  thrpt    5   1014403.744 ±  44355.536   ops/s
[info] MissingRequiredFieldsReading.jsoniterScalaWithoutDump      N/A  thrpt    5  11935097.222 ± 157861.180   ops/s
[info] MutableBitSetReading.jsoniterScala                         128  thrpt    5   1997689.247 ±  18351.717   ops/s
[info] MutableLongMapOfBooleansReading.jsoniterScala              128  thrpt    5    429808.917 ±   9082.678   ops/s
[info] MutableMapOfIntsToBooleansReading.jsoniterScala            128  thrpt    5    473097.845 ±  13637.323   ops/s
[info] MutableSetOfIntsReading.jsoniterScala                      128  thrpt    5    566357.815 ±  12396.315   ops/s
[info] NestedStructsReading.jsoniterScala                         128  thrpt    5    560310.042 ±  12683.813   ops/s
[info] OpenRTBReading.jsoniterScala                               N/A  thrpt    5    375201.115 ±   2892.235   ops/s
[info] PrimitivesReading.jsoniterScala                            N/A  thrpt    5   9636541.731 ±  60582.585   ops/s
[info] SetOfIntsReading.jsoniterScala                             128  thrpt    5    273997.579 ±   5638.281   ops/s
[info] StringOfAsciiCharsReading.jsoniterScala                    128  thrpt    5  10024396.316 ± 300910.621   ops/s
[info] StringOfEscapedCharsReading.jsoniterScala                  128  thrpt    5   3151768.387 ±  54043.885   ops/s
[info] StringOfNonAsciiCharsReading.jsoniterScala                 128  thrpt    5   3590377.573 ±  49702.075   ops/s
[info] TwitterAPIReading.jsoniterScala                            N/A  thrpt    5     57575.812 ±   1531.197   ops/s
[info] VectorOfBooleansReading.jsoniterScala                      128  thrpt    5   2588563.723 ±  16079.294   ops/s
```
## After
```
[info] Benchmark                                               (size)   Mode  Cnt         Score         Error   Units
[info] ADTReading.jsoniterScala                                   N/A  thrpt    5   6098381.091 ±  266409.259   ops/s
[info] AnyValsReading.jsoniterScala                               N/A  thrpt    5   8643265.108 ±  119079.434   ops/s
[info] ArrayBufferOfBooleansReading.jsoniterScala                 128  thrpt    5   2151282.322 ±   58182.061   ops/s
[info] ArrayOfBigDecimalsReading.jsoniterScala                    128  thrpt    5    133179.483 ±    1235.737   ops/s
[info] ArrayOfBigIntsReading.jsoniterScala                        128  thrpt    5    227879.055 ±    5509.348   ops/s
[info] ArrayOfBooleansReading.jsoniterScala                       128  thrpt    5   2862539.970 ±   89024.495   ops/s
[info] ArrayOfBytesReading.jsoniterScala                          128  thrpt    5   1679502.227 ±   20180.209   ops/s
[info] ArrayOfCharsReading.jsoniterScala                          128  thrpt    5   1835618.448 ±   20472.310   ops/s
[info] ArrayOfDoublesReading.jsoniterScala                        128  thrpt    5    387785.772 ±    2711.145   ops/s
[info] ArrayOfDurationsReading.jsoniterScala                      128  thrpt    5    288386.987 ±    1478.457   ops/s
[info] ArrayOfEnumADTsReading.jsoniterScala                       128  thrpt    5    766228.541 ±    6789.961   ops/s
[info] ArrayOfEnumsReading.jsoniterScala                          128  thrpt    5    424391.761 ±    3516.945   ops/s
[info] ArrayOfFloatsReading.jsoniterScala                         128  thrpt    5    528132.925 ±    2932.756   ops/s
[info] ArrayOfInstantsReading.jsoniterScala                       128  thrpt    5    466328.358 ±    3126.353   ops/s
[info] ArrayOfIntsReading.jsoniterScala                           128  thrpt    5   1011803.848 ±    9821.458   ops/s
[info] ArrayOfJavaEnumsReading.jsoniterScala                      128  thrpt    5    834539.523 ±   21107.675   ops/s
[info] ArrayOfLocalDateTimesReading.jsoniterScala                 128  thrpt    5    434741.830 ±    8802.643   ops/s
[info] ArrayOfLocalDatesReading.jsoniterScala                     128  thrpt    5    943373.662 ±   11663.910   ops/s
[info] ArrayOfLocalTimesReading.jsoniterScala                     128  thrpt    5    657836.081 ±    2681.980   ops/s
[info] ArrayOfLongsReading.jsoniterScala                          128  thrpt    5    682729.908 ±   32851.600   ops/s
[info] ArrayOfMonthDaysReading.jsoniterScala                      128  thrpt    5   1214986.487 ±    8279.412   ops/s
[info] ArrayOfOffsetDateTimesReading.jsoniterScala                128  thrpt    5    367528.711 ±   16234.175   ops/s
[info] ArrayOfOffsetTimesReading.jsoniterScala                    128  thrpt    5    518010.347 ±    2624.850   ops/s
[info] ArrayOfPeriodsReading.jsoniterScala                        128  thrpt    5    325227.109 ±    1948.376   ops/s
[info] ArrayOfShortsReading.jsoniterScala                         128  thrpt    5   1362706.002 ±   13229.047   ops/s
[info] ArrayOfUUIDsReading.jsoniterScala                          128  thrpt    5    576527.855 ±   11333.032   ops/s
[info] ArrayOfYearMonthsReading.jsoniterScala                     128  thrpt    5   1376451.613 ±   10282.236   ops/s
[info] ArrayOfYearsReading.jsoniterScala                          128  thrpt    5   1604094.302 ±   34809.483   ops/s
[info] ArrayOfZoneIdsReading.jsoniterScala                        128  thrpt    5    375614.146 ±    5416.154   ops/s
[info] ArrayOfZoneOffsetsReading.jsoniterScala                    128  thrpt    5   1426479.379 ±   15168.046   ops/s
[info] ArrayOfZonedDateTimesReading.jsoniterScala                 128  thrpt    5    152691.150 ±    2945.818   ops/s
[info] ArraySeqOfBooleansReading.jsoniterScala                    128  thrpt    5   2830168.756 ±  160647.816   ops/s
[info] Base16Reading.jsoniterScala                                128  thrpt    5   8743571.240 ±  451788.631   ops/s
[info] Base64Reading.jsoniterScala                                128  thrpt    5   6559664.823 ±  110559.086   ops/s
[info] BigDecimalReading.jsoniterScala                            128  thrpt    5   6905781.135 ±  528764.541   ops/s
[info] BigIntReading.jsoniterScala                                128  thrpt    5   6866695.413 ±   57573.454   ops/s
[info] BitSetReading.jsoniterScala                                128  thrpt    5   1559740.732 ±   97987.778   ops/s
[info] ExtractFieldsReading.jsoniterScala                         128  thrpt    5    252758.557 ±     561.015   ops/s
[info] GeoJSONReading.jsoniterScala                               N/A  thrpt    5     70181.014 ±    1199.698   ops/s
[info] GitHubActionsAPIReading.jsoniterScala                      N/A  thrpt    5    833443.474 ±    1456.829   ops/s
[info] GoogleMapsAPIReading.jsoniterScala                         N/A  thrpt    5     36712.374 ±     137.620   ops/s
[info] IntMapOfBooleansReading.jsoniterScala                      128  thrpt    5    271422.336 ±   13163.982   ops/s
[info] IntReading.jsoniterScala                                   N/A  thrpt    5  83922855.833 ± 2553163.618   ops/s
[info] ListOfBooleansReading.jsoniterScala                        128  thrpt    5   2211277.292 ±   12813.047   ops/s
[info] MapOfIntsToBooleansReading.jsoniterScala                   128  thrpt    5    217285.484 ±    1226.142   ops/s
[info] MissingRequiredFieldsReading.jsoniterScala                 N/A  thrpt    5   6177488.618 ±  186589.687   ops/s
[info] MissingRequiredFieldsReading.jsoniterScalaWithStacktrace   N/A  thrpt    5    965627.660 ±   71012.325   ops/s
[info] MissingRequiredFieldsReading.jsoniterScalaWithoutDump      N/A  thrpt    5  12181174.260 ±  325438.146   ops/s
[info] MutableBitSetReading.jsoniterScala                         128  thrpt    5   1892618.441 ±   13638.911   ops/s
[info] MutableLongMapOfBooleansReading.jsoniterScala              128  thrpt    5    421703.115 ±    1274.617   ops/s
[info] MutableMapOfIntsToBooleansReading.jsoniterScala            128  thrpt    5    475510.513 ±   23248.362   ops/s
[info] MutableSetOfIntsReading.jsoniterScala                      128  thrpt    5    572593.860 ±   10135.111   ops/s
[info] NestedStructsReading.jsoniterScala                         128  thrpt    5    526446.788 ±    6066.700   ops/s
[info] OpenRTBReading.jsoniterScala                               N/A  thrpt    5    354609.378 ±    2045.059   ops/s
[info] PrimitivesReading.jsoniterScala                            N/A  thrpt    5   9408659.991 ±  221388.614   ops/s
[info] SetOfIntsReading.jsoniterScala                             128  thrpt    5    272880.392 ±    1010.124   ops/s
[info] StringOfAsciiCharsReading.jsoniterScala                    128  thrpt    5  15655962.746 ±   83237.013   ops/s
[info] StringOfEscapedCharsReading.jsoniterScala                  128  thrpt    5   3246458.535 ±   17403.505   ops/s
[info] StringOfNonAsciiCharsReading.jsoniterScala                 128  thrpt    5   3555793.098 ±   21413.789   ops/s
[info] TwitterAPIReading.jsoniterScala                            N/A  thrpt    5     57864.902 ±     326.815   ops/s
[info] VectorOfBooleansReading.jsoniterScala                      128  thrpt    5   2298477.021 ±   19594.256   ops/s
```