@plokhotnyuk 2022-08-21
## After using LUT for checking of whitespaces
```
[info] Benchmark                                               (size)   Mode  Cnt         Score        Error   Units
[info] ADTReading.jsoniterScala                                   N/A  thrpt    5   6205436.690 ±  43500.778   ops/s
[info] AnyValsReading.jsoniterScala                               N/A  thrpt    5  10151192.101 ± 176410.702   ops/s
[info] ArrayBufferOfBooleansReading.jsoniterScala                 128  thrpt    5   2024344.100 ±  18934.680   ops/s
[info] ArrayOfBigDecimalsReading.jsoniterScala                    128  thrpt    5    150053.500 ±    412.746   ops/s
[info] ArrayOfBigIntsReading.jsoniterScala                        128  thrpt    5    219038.338 ±  24167.402   ops/s
[info] ArrayOfBooleansReading.jsoniterScala                       128  thrpt    5   2600803.553 ±  36946.009   ops/s
[info] ArrayOfBytesReading.jsoniterScala                          128  thrpt    5   1648230.404 ±   5817.290   ops/s
[info] ArrayOfCharsReading.jsoniterScala                          128  thrpt    5   1863350.437 ±  17074.017   ops/s
[info] ArrayOfDoublesReading.jsoniterScala                        128  thrpt    5    386987.241 ±  12429.513   ops/s
[info] ArrayOfDurationsReading.jsoniterScala                      128  thrpt    5    279670.019 ±   5028.036   ops/s
[info] ArrayOfEnumADTsReading.jsoniterScala                       128  thrpt    5    774657.574 ±   4572.800   ops/s
[info] ArrayOfEnumsReading.jsoniterScala                          128  thrpt    5    424593.466 ±   7417.506   ops/s
[info] ArrayOfFloatsReading.jsoniterScala                         128  thrpt    5    531606.009 ±  13775.456   ops/s
[info] ArrayOfInstantsReading.jsoniterScala                       128  thrpt    5    453430.513 ±  13279.742   ops/s
[info] ArrayOfIntsReading.jsoniterScala                           128  thrpt    5   1019789.281 ±  27682.263   ops/s
[info] ArrayOfJavaEnumsReading.jsoniterScala                      128  thrpt    5    843613.778 ±   5152.636   ops/s
[info] ArrayOfLocalDateTimesReading.jsoniterScala                 128  thrpt    5    428117.225 ±   4789.188   ops/s
[info] ArrayOfLocalDatesReading.jsoniterScala                     128  thrpt    5    957599.266 ±  16684.419   ops/s
[info] ArrayOfLocalTimesReading.jsoniterScala                     128  thrpt    5    669135.544 ±   8086.645   ops/s
[info] ArrayOfLongsReading.jsoniterScala                          128  thrpt    5    698307.220 ±   9688.494   ops/s
[info] ArrayOfMonthDaysReading.jsoniterScala                      128  thrpt    5   1257862.893 ±  36979.230   ops/s
[info] ArrayOfOffsetDateTimesReading.jsoniterScala                128  thrpt    5    372011.946 ±   2100.847   ops/s
[info] ArrayOfOffsetTimesReading.jsoniterScala                    128  thrpt    5    532923.106 ±  14579.294   ops/s
[info] ArrayOfPeriodsReading.jsoniterScala                        128  thrpt    5    327280.773 ±  13119.998   ops/s
[info] ArrayOfShortsReading.jsoniterScala                         128  thrpt    5   1414447.627 ±  38932.491   ops/s
[info] ArrayOfUUIDsReading.jsoniterScala                          128  thrpt    5    574258.694 ±   6324.537   ops/s
[info] ArrayOfYearMonthsReading.jsoniterScala                     128  thrpt    5   1394453.994 ±  30732.349   ops/s
[info] ArrayOfYearsReading.jsoniterScala                          128  thrpt    5   1867392.097 ±   7431.133   ops/s
[info] ArrayOfZoneIdsReading.jsoniterScala                        128  thrpt    5    382485.654 ±   4962.083   ops/s
[info] ArrayOfZoneOffsetsReading.jsoniterScala                    128  thrpt    5   1540590.842 ±  13486.670   ops/s
[info] ArrayOfZonedDateTimesReading.jsoniterScala                 128  thrpt    5    152852.611 ±    512.915   ops/s
[info] ArraySeqOfBooleansReading.jsoniterScala                    128  thrpt    5   2558069.217 ±  79253.612   ops/s
[info] Base16Reading.jsoniterScala                                128  thrpt    5   8810033.041 ±  13884.867   ops/s
[info] Base64Reading.jsoniterScala                                128  thrpt    5   6558193.937 ±  98827.511   ops/s
[info] BigDecimalReading.jsoniterScala                            128  thrpt    5   6940886.464 ±  62707.261   ops/s
[info] BigIntReading.jsoniterScala                                128  thrpt    5   6733620.469 ±  54908.285   ops/s
[info] BitSetReading.jsoniterScala                                128  thrpt    5   1499892.111 ±   8062.780   ops/s
[info] ExtractFieldsReading.jsoniterScala                         128  thrpt    5    247146.212 ±   1326.321   ops/s
[info] GeoJSONReading.jsoniterScala                               N/A  thrpt    5     70806.922 ±    546.085   ops/s
[info] GitHubActionsAPIReading.jsoniterScala                      N/A  thrpt    5    745858.491 ±   8471.847   ops/s
[info] GoogleMapsAPIReading.jsoniterScala                         N/A  thrpt    5     31423.282 ±     92.905   ops/s
[info] IntMapOfBooleansReading.jsoniterScala                      128  thrpt    5    264678.945 ±  16029.331   ops/s
[info] IntReading.jsoniterScala                                   N/A  thrpt    5  88423891.407 ±  89464.926   ops/s
[info] ListOfBooleansReading.jsoniterScala                        128  thrpt    5   2088679.300 ±  38404.928   ops/s
[info] MapOfIntsToBooleansReading.jsoniterScala                   128  thrpt    5    210267.975 ±   5054.269   ops/s
[info] MissingRequiredFieldsReading.jsoniterScala                 N/A  thrpt    5   6101509.407 ±  19084.998   ops/s
[info] MissingRequiredFieldsReading.jsoniterScalaWithStacktrace   N/A  thrpt    5    966203.541 ±  42964.819   ops/s
[info] MissingRequiredFieldsReading.jsoniterScalaWithoutDump      N/A  thrpt    5  11934337.817 ± 467002.229   ops/s
[info] MutableBitSetReading.jsoniterScala                         128  thrpt    5   1957723.461 ±  92632.928   ops/s
[info] MutableLongMapOfBooleansReading.jsoniterScala              128  thrpt    5    419723.257 ±  15212.106   ops/s
[info] MutableMapOfIntsToBooleansReading.jsoniterScala            128  thrpt    5    478968.861 ±   3956.110   ops/s
[info] MutableSetOfIntsReading.jsoniterScala                      128  thrpt    5    593881.430 ±   4724.883   ops/s
[info] NestedStructsReading.jsoniterScala                         128  thrpt    5    546496.741 ±  15162.154   ops/s
[info] OpenRTBReading.jsoniterScala                               N/A  thrpt    5    365313.173 ±   3237.313   ops/s
[info] PrimitivesReading.jsoniterScala                            N/A  thrpt    5  10331826.211 ±  99624.854   ops/s
[info] SetOfIntsReading.jsoniterScala                             128  thrpt    5    261761.041 ±   2693.155   ops/s
[info] StringOfAsciiCharsReading.jsoniterScala                    128  thrpt    5  10020639.286 ± 204394.743   ops/s
[info] StringOfEscapedCharsReading.jsoniterScala                  128  thrpt    5   3274481.927 ±  13783.982   ops/s
[info] StringOfNonAsciiCharsReading.jsoniterScala                 128  thrpt    5   3514675.382 ±  87203.374   ops/s
[info] TwitterAPIReading.jsoniterScala                            N/A  thrpt    5     53386.013 ±   2175.792   ops/s
[info] VectorOfBooleansReading.jsoniterScala                      128  thrpt    5   2563047.414 ±  13930.703   ops/s
```